+++ lib/create.dart
@@ -55,11 +55,12 @@ PORT = 3000
     ''', mode: FileMode.append);
   }
 
-  File main = File('$directory/bin/main.dart');
+  Directory('$directory/lib').createSync();
+  Directory('$directory/public').createSync();
+  File('$directory/public/favicon.ico').createSync();
 
-  if (!main.existsSync()) {
-    main.createSync();
-  }
+  File main = File('$directory/lib/main.dart');
+  main.createSync();
 
   main.writeAsStringSync('''
 import 'package:lucifer/lucifer.dart';
@@ -69,6 +70,7 @@ void main() async {
   final port = env('PORT') ?? 3000;
 
   app.use(logger());
+  app.use(static('public'));
 
   app.get('/', (Req req, Res res) async {
     await res.send('Hello Detective');
@@ -81,11 +83,19 @@ void main() async {
 }
   ''');
 
-  File projectFile = File('$directory/bin/$project.dart');
+  File bin = File('$directory/bin/main.dart');
+  bin.createSync();
 
-  if (projectFile.existsSync()) {
-    projectFile.deleteSync();
-  }
+  bin.writeAsStringSync('''
+import 'package:$project/main.dart' as $project;
+
+void main() {
+  $project.main();
+}
+  ''');
+
+  File proj = File('$directory/bin/$project.dart');
+  proj.deleteSync();
 
   shell = shell.pushd(project);
   await shell.run('pub get');