+++ lib/lucy.dart
@@ -1,3 +1,4 @@
+export 'build.dart';
 export 'controller.dart';
 export 'create.dart';
 export 'db.dart';